requirements.txt
django==5.2.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
pillow==11.2.1
gunicorn==23.0.0
